Separated parts of the potluck view into separate components. Added check for admin and display admin sections accordingly
