+++ potluck-planner/src/components/Potluck/PotluckView/PotluckView.scss
@@ -1,0 +1,3 @@
+.non-admin {
+	display: none;
+}
